[Refactor] Extract assert_response
